requirements.txt
langchain 
wikipedia 
langchain-community 
langchain-huggingface
python-dotenv
pydantic
duckduckgo-search
transformers
torch